(weeks projects & practices)

+++ week-3/practice_6/src/main.rs
@@ -1,5 +1,5 @@
 fn main() {
-    let interest:f32 = 8;  // integer assigned to float variable
+    let interest:i32 = 8;  // integer assigned to float variable
 
     println!(" interest is {} ", interest);
 }
